live-api/backend/requirements.txt
websockets==15.0.1
asyncio==3.4.3
google-auth==2.40.1
fastapi==0.115.12
uvicorn==0.34.2
gunicorn==23.0.0
requests==2.28.1
